change of type double -> float in the variables of the entities area and ox

+++ src/main/java/com/jetbov/exercice/core/entities/Area.java
@@ -2,6 +2,6 @@ package com.jetbov.exercice.core.entities;
 
 public record Area(
         String name,
-        double gmd,
+        float gmd,
         int maxCapacity
 ) {}
